major changes
imported node.js and socket.io for when the multiplayer capabilities are added. Created a highlighting system to highlight possible moves when you click on a piece. Highlight has only been implemented for pawns and rooks at this point. To Do: Finish other pieces highlight, fix issue that you cant choose to move to a spot where another piece is, add turn system and movePiece function.